requirements.txt
# Project 3: Image Caption Generator
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
requests>=2.31.0
python-multipart>=0.0.6
Pillow>=10.0.0